% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{thicknessGlabTop2}
\alias{thicknessGlabTop2}
\title{glabTop2 model for glacier thickness estimation}
\usage{
thicknessGlabTop2(
  DEM,
  glacier.outline,
  g,
  rho,
  h.min,
  rc,
  f.range,
  n,
  idw.P,
  hga,
  ncores
)
}
\arguments{
\item{DEM}{glacier DEM (Formal class RasterLayer).Must be projected}

\item{glacier.outline}{glacier boundary shapefile (Formal class SpatialPolygons).Must have same projection as the DEM}

\item{g}{gravitational acceleration in ms-2 [default 9.8]}

\item{rho}{density of ice in kgm-3[default 900]}

\item{h.min}{minimum elevation difference(in m) in the buffer formed around randomly selected cell, for calculation of mean surface slope[default 50]}

\item{rc}{fraction of randomly selected cells from glacier inner cells to be used for calculation of thickness[default 0.3]}

\item{f.range}{range of values of shape factors 'f', for which thickness is to be estimated [default seq(0.6,0.9,0.01)]}

\item{n}{number of thickness estimation for a particular value of shape factor 'f' [default n=3]}

\item{idw.P}{power of the IDW(inverse distant weighted) Interpolation [default 3]}

\item{hga}{value assigned to glacier adjacent cells [default 15]}

\item{ncores}{number of cores of your CPU, to be used for running the code [default: total number of codes of CPU-3]}
}
\value{
Returns a list of two objects -
          (i)'averageThickness'  :  (Formal Class RasterLayer) average of the thickness estimated, using the selected values of shape factor 'f'
          (ii)'thickness.shapeFactors'  : (List of Formal class RasterLayer)thickness estimated for each
}
\description{
Estimation of glacier thickness, based on the Glacier Bed Topography (GlabTop2) model (Frey et al., 2014)
}
\examples{
#creating a sample glacier DEM
proj<-"+proj=utm +zone=44N +datum=WGS84 +units=m +no_defs "
values<-round(runif(2793,4000,6000),digits = 2)
r.extent<-raster::extent(308004.8, 336707.3, 3399079, 3423596)
r<-raster::raster(ext=r.extent,resolution=500, crs=proj,vals=values)
#creating a sample glacier boundary
shp <- as(raster::extent(309004.8, 334707.3, 3400579, 3420596), "SpatialPolygons")
sp::proj4string(shp) <- "+proj=utm +zone=44N +datum=WGS84 +units=m +no_defs"

#estimating glacier thickness
thickness<-thicknessGlabTop2(DEM=r,glacier.outline=shp,f.range=c(0.65,0.66),n=1,ncores=2)

}
\references{
RAAJ Ramsankaran, Ankur Pandit & Mohd Farooq Azam (2018) Spatially distributed ice-thickness modelling for Chhota Shigri Glacier in western Himalayas, India, International Journal of Remote Sensing, 39:10, 3320-3343, DOI: 10.1080/01431161.2018.1441563

Frey, H., Machguth, H., Huss, M., Huggel, C., Bajracharya, S., Bolch, T., Kulkarni, A., Linsbauer, A., Salzmann, N., and Stoffel, M.: Estimating the volume of glaciers in the Himalayan–Karakoram region using different methods, The Cryosphere, 8, 2313–2333, https://doi.org/10.5194/tc-8-2313-2014, 2014.
}
\author{
Aniruddha
}
